+++ app/config.py
@@ -1,3 +1,5 @@
+import os
+
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 
@@ -9,7 +11,10 @@ class Settings(BaseSettings):
     REDDIT_USER_AGENT: str = "testscript by echobot392"
     REDDIT_BASE_URL: str = "https://www.reddit.com/"
 
-    model_config = SettingsConfigDict(env_file='.env', extra='ignore')
+    BASEDIR: str = os.path.abspath(os.path.dirname(__file__))
+    ENV_FILE_PATH: str = os.path.join(BASEDIR, '..', '.env')
+
+    model_config = SettingsConfigDict(env_file=ENV_FILE_PATH, extra='ignore')
 
 
 settings = Settings()
