+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 pythonpath = . app
-addopts = -s -v --disable-pytest-warnings --cov=app --cov-report=html --cov-branch --cov-fail-under=30 .
+addopts = -s -v --disable-pytest-warnings --cov=app --cov-report=html --cov-branch --cov-fail-under=30
 asyncio_mode = auto
